# Eliminate tuple allocations in seekReadCustomAttributeRow via a custom reader for attribute rows