This PR is more for folks like @dsyme and @TIHan to look at and see what they think.

From this issue: https://developercommunity.visualstudio.com/content/problem/1035124/trace-of-editing-experience-with-in-memory-cross-p-1.html

The submitted trace showed 120MB of allocations coming from this function, most of which are 3-tuples (~5.7% of all allocations in the sample):

![image](https://user-images.githubusercontent.com/6309070/82162392-ee40f580-9858-11ea-9073-31bff4687a06.png)

It's also 0.7% of total CPU time:

![image](https://user-images.githubusercontent.com/6309070/82162408-03b61f80-9859-11ea-817a-497da00e0f81.png)

Since the data being generated is all small value types, maybe a struct tuple would be better here.

Looking through what all calls this and the frequency, it looks like it's just a part of normal compiler routines so if we think this would speed things up a bit then that'd be nice.