@TIHan 2020-05-18
I think we should fix this despite not pushing the SRM PR. 

I don't think we can switch to a SRM implementation because of the risk involved. We need to have a huge suite of tests that directly test our IL metadata reader.