@cartermp 2020-05-18
Yeah, #4716 would be good. I just scoped this (and some other PRs) to address only one specific thing to keep them as small as possible rather than go through and convert a lot of the obvious ones to use structs (like this little chungus: https://github.com/dotnet/fsharp/blob/master/src/absil/ilread.fs#L256)