@cartermp 2020-05-18
I took a trace from using VS after editing the compiler codebase for an hour and this is basically near the top in terms of allocations

![image](https://user-images.githubusercontent.com/6309070/82241474-1419d900-98f1-11ea-9a24-1869d1806785.png)

While #4716 would also improve things this is a bit more targeted. @TIHan is there a good set of benchmarks we could potentially use here? I'd like to see if switching to a string tuple does help since it's shown up in two separate traces as a top thing