@TIHan 2020-05-26
@cartermp You could try using this benchmark. https://github.com/dotnet/fsharp/blob/master/benchmarks/CompilerServiceBenchmarks/Program.fs#L194

Maybe tweak it to read all the attributes in an assembly?