@cartermp 2020-05-29
This one was a little iffy for me (`parentIdx` and `typeIdx` are a struct of two ints), hence the original message. As expected, allocations are reduced noticeably, but CPU time did go up slightly in ILReading.

``` ini

BenchmarkDotNet=v0.11.3, OS=Windows 10.0.19041
Intel Core i7-6700K CPU 4.00GHz (Skylake), 1 CPU, 8 logical and 4 physical cores
  [Host]     : .NET Framework 4.7.2 (CLR 4.0.30319.42000), 64bit LegacyJIT-v4.8.4180.0 DEBUG
  Job-FEFVFA : .NET Framework 4.7.2 (CLR 4.0.30319.42000), 64bit RyuJIT-v4.8.4180.0

InvocationCount=1  UnrollFactor=1  

```
Before:
|                                 Method |        Mean |     Error |    StdDev |       Median | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|--------------------------------------- |------------:|----------:|----------:|-------------:|------------:|------------:|------------:|--------------------:|
|                   ParsingTypeCheckerFs |   191.39 ms |  1.893 ms |  1.678 ms |   190.873 ms |   7000.0000 |   2000.0000 |           - |            42.77 MB |
|                              ILReading | 1,529.46 ms | 30.370 ms | 49.041 ms | 1,527.549 ms | 191000.0000 |  63000.0000 |   6000.0000 |          1112.69 MB |
| TypeCheckFileWith100ReferencedProjects |    13.11 ms |  1.543 ms |  4.550 ms |     9.731 ms |   1000.0000 |   1000.0000 |           - |             7.38 MB |

After:
|                                 Method |         Mean |      Error |     StdDev | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|--------------------------------------- |-------------:|-----------:|-----------:|------------:|------------:|------------:|--------------------:|
|                   ParsingTypeCheckerFs |   206.865 ms |  3.6175 ms |  3.3838 ms |   7000.0000 |   2000.0000 |           - |            42.77 MB |
|                              ILReading | 1,585.444 ms | 11.3409 ms | 10.0534 ms | 174000.0000 |  57000.0000 |   5000.0000 |          1037.59 MB |
| TypeCheckFileWith100ReferencedProjects |     9.480 ms |  0.1684 ms |  0.1493 ms |   1000.0000 |   1000.0000 |           - |              7.4 MB |