@cartermp 2020-06-27
@TIHan what are your thoughts here? I think 17MB less allocations per operation outweighs the slight increase in CPU time. It's also worth noting that the measured CPU time now also has less variance than before.