@cartermp 2020-08-05
Okay, I tried this with @TIHan's proposed approach in #4716 and these were the results of the compiler service benchmarks:

``` ini

BenchmarkDotNet=v0.11.3, OS=Windows 10.0.19041
Intel Core i7-6700K CPU 4.00GHz (Skylake), 1 CPU, 8 logical and 4 physical cores
  [Host]     : .NET Framework 4.7.2 (CLR 4.0.30319.42000), 64bit LegacyJIT-v4.8.4180.0 DEBUG
  Job-FEFVFA : .NET Framework 4.7.2 (CLR 4.0.30319.42000), 64bit RyuJIT-v4.8.4180.0

InvocationCount=1  UnrollFactor=1  

```
|                                 Method |        Mean |     Error |    StdDev |       Median | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|--------------------------------------- |------------:|----------:|----------:|-------------:|------------:|------------:|------------:|--------------------:|
|                   ParsingTypeCheckerFs |   193.71 ms |  1.846 ms |  1.726 ms |   193.127 ms |   6000.0000 |   2000.0000 |           - |            40.41 MB |
|                              ILReading | 1,485.79 ms | 26.141 ms | 24.452 ms | 1,486.826 ms | 149000.0000 |  51000.0000 |   6000.0000 |           899.01 MB |
| TypeCheckFileWith100ReferencedProjects |    11.73 ms |  1.503 ms |  4.431 ms |     8.845 ms |   1000.0000 |   1000.0000 |           - |             7.34 MB |

CPU time is improved over the current and new implementation by ~2%.

Memory usage per op is 19% better than the current implementation, and 13% better than the change I made.

So that's what I'll go with.