@cartermp 2020-08-05
Might be worth trying to unify on this? https://github.com/dotnet/fsharp/blob/master/src/absil/ilread.fs#L746