@cartermp 2020-08-05
I wont bother trying to consolidate with https://github.com/dotnet/fsharp/blob/master/src/absil/ilread.fs#L746

It's a bit of additional churn and I haven't seen evidence from trace data that the tuples the functions that call this thing have an impact.