@abelbraaksma 2020-08-05
> InvocationCount=1  UnrollFactor=1  

The numbers look intriguing, but with a single invocation, I'm not sure they proof anything (and may be better, or worse, than what you're seeing). Maybe you ran into a bug of BDN that I reported there (and isn't fixed yet), which chooses 1 invocation when there's a lot secundairy jitting going on.

Can you share the BDN code, or force it with higher invocation count?